Join user and posts

+++ salfny/salfny/src/main/java/com/swe/salfny/User/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface UserRepository extends JpaRepository<UserData, Long> {
+public interface UserRepository extends JpaRepository<UserData, Integer> {
     @Query("SELECT u FROM UserData u WHERE u.email = ?1")
     public UserData findByEmail(String email);
